refactor: #50 StudyHeader index 파일 분리
